Break snare.py
