+++ snare.py
@@ -16,24 +16,20 @@ GNU General Public License for more details.
 import argparse
 import asyncio
 import configparser
-import grp
 import json
 import mimetypes
 import multiprocessing
 import os
-import pwd
 import sys
 import time
 import uuid
 from concurrent.futures import ProcessPoolExecutor
 from urllib.parse import urlparse, unquote, parse_qsl
-from versions_manager import VersionManager
 import aiohttp
 import git
 import pip
 import re
 import logging
-import logger
 
 try:
     from aiohttp.web import StaticResource as StaticRoute
@@ -43,142 +39,13 @@ except ImportError:
 from bs4 import BeautifulSoup
 import cssutils
 import netifaces as ni
-from converter import Converter
 from server import HttpRequestHandler
-
-
-def create_initial_config():
-    cfg = configparser.ConfigParser()
-    cfg['WEB-TOOLS'] = dict(google='', bing='')
-    with open('/opt/snare/snare.cfg', 'w') as configfile:
-        cfg.write(configfile)
-
-
-def snare_setup():
-    if os.getuid() != 0:
-        print('Snare has to be started as root!')
-        sys.exit(1)
-    # Create folders
-    if not os.path.exists('/opt/snare'):
-        os.mkdir('/opt/snare')
-    if not os.path.exists('/opt/snare/pages'):
-        os.mkdir('/opt/snare/pages')
-    # Write pid to pid file
-    with open('/opt/snare/snare.pid', 'wb') as pid_fh:
-        pid_fh.write(str(os.getpid()).encode('utf-8'))
-    # Config file
-    if not os.path.exists('/opt/snare/snare.cfg'):
-        create_initial_config()
-    # Read or create the sensor id
-    uuid_file_path = '/opt/snare/snare.uuid'
-    if os.path.exists(uuid_file_path):
-        with open(uuid_file_path, 'rb') as uuid_fh:
-            snare_uuid = uuid_fh.read()
-        return snare_uuid
-    else:
-        with open(uuid_file_path, 'wb') as uuid_fh:
-            snare_uuid = str(uuid.uuid4()).encode('utf-8')
-            uuid_fh.write(snare_uuid)
-        return snare_uuid
-
-
-def drop_privileges():
-    uid_name = 'nobody'
-    wanted_user = pwd.getpwnam(uid_name)
-    gid_name = grp.getgrgid(wanted_user.pw_gid).gr_name
-    wanted_group = grp.getgrnam(gid_name)
-    os.setgid(wanted_group.gr_gid)
-    os.setuid(wanted_user.pw_uid)
-    new_user = pwd.getpwuid(os.getuid())
-    new_group = grp.getgrgid(os.getgid())
-    print('privileges dropped, running as "{}:{}"'.format(new_user.pw_name, new_group.gr_name))
-
-
-def add_meta_tag(page_dir, index_page):
-    google_content = config['WEB-TOOLS']['google']
-    bing_content = config['WEB-TOOLS']['bing']
-
-    if not google_content and not bing_content:
-        return
-
-    main_page_path = os.path.join('/opt/snare/pages/', page_dir, index_page)
-    with open(main_page_path) as main:
-        main_page = main.read()
-    soup = BeautifulSoup(main_page, 'html.parser')
-
-    if (google_content and soup.find("meta", attrs={"name": "google-site-verification"}) is None):
-        google_meta = soup.new_tag('meta')
-        google_meta.attrs['name'] = 'google-site-verification'
-        google_meta.attrs['content'] = google_content
-        soup.head.append(google_meta)
-    if (bing_content and soup.find("meta", attrs={"name": "msvalidate.01"}) is None):
-        bing_meta = soup.new_tag('meta')
-        bing_meta.attrs['name'] = 'msvalidate.01'
-        bing_meta.attrs['content'] = bing_content
-        soup.head.append(bing_meta)
-
-    html = soup.prettify("utf-8")
-    with open(main_page_path, "wb") as file:
-        file.write(html)
-
-
-def compare_version_info(timeout):
-    while True:
-        repo = git.Repo(os.getcwd())
-        try:
-            rem = repo.remote()
-            res = rem.fetch()
-            diff_list = res[0].commit.diff(repo.heads.master)
-        except TimeoutError:
-            print('timeout fetching the repository version')
-        else:
-            if diff_list:
-                print('you are running an outdated version, SNARE will be updated and restarted')
-                repo.git.reset('--hard')
-                repo.heads.master.checkout()
-                repo.git.clean('-xdf')
-                repo.remotes.origin.pull()
-                pip.main(['install', '-r', 'requirements.txt'])
-                os.execv(sys.executable, [sys.executable, __file__] + sys.argv[1:])
-                return
-            else:
-                print('you are running the latest version')
-            time.sleep(timeout)
-
-
-def parse_timeout(timeout):
-    result = None
-    timeouts_coeff = {
-        'M': 60,
-        'H': 3600,
-        'D': 86400
-    }
-
-    form = timeout[-1]
-    if form not in timeouts_coeff.keys():
-        print('Bad timeout format, default will be used')
-        parse_timeout('24H')
-    else:
-        result = int(timeout[:-1])
-        result *= timeouts_coeff[form]
-    return result
-
-
-async def check_tanner():
-    vm = VersionManager()
-    async with aiohttp.ClientSession() as client:
-        req_url = 'http://{}:8090/version'.format(args.tanner)
-        try:
-            resp = await client.get(req_url)
-            result = await resp.json()
-            version = result["version"]
-            vm.check_compatibility(version)
-        except aiohttp.ClientOSError:
-            print("Can't connect to tanner host {}".format(req_url))
-            exit(1)
-        else:
-            await resp.release()
-
+from startup import StartUp
+from utils.logger import Logger
+from utils.tag_adder import add_meta_tag
+from utils.timeout_parser import parse_timeout
+from utils.converter import Converter
+from utils.versions_manager import VersionManager
 
 if __name__ == '__main__':
     print(r"""
@@ -189,7 +56,6 @@ if __name__ == '__main__':
 /____/_/ |_/_/  |_/_/ |_/_____/
 
     """)
-    snare_uuid = snare_setup()
     parser = argparse.ArgumentParser()
     page_group = parser.add_mutually_exclusive_group(required=True)
     page_group.add_argument("--page-dir", help="name of the folder to be served")
@@ -213,12 +79,14 @@ if __name__ == '__main__':
     args = parser.parse_args()
     base_path = '/opt/snare/'
     base_page_path = '/opt/snare/pages/'
+    startup = StartUp(args)
+    snare_uuid = startup.snare_setup()
     config = configparser.ConfigParser()
     config.read(os.path.join(base_path, args.config))
     
     log_debug = args.log_dir + "snare.log"
     log_err = args.log_dir + "snare.err"      
-    logger.Logger.create_logger(log_debug, log_err, __package__)
+    Logger.create_logger(log_debug, log_err, __package__)
 
     if args.list_pages:
         print('Available pages:\n')
@@ -243,21 +111,21 @@ if __name__ == '__main__':
                                        os.path.join(meta_info[args.index_page]['hash']))):
         print('can\'t create meta tag')
     else:
-        add_meta_tag(args.page_dir, meta_info[args.index_page]['hash'])
+        add_meta_tag(args.page_dir, meta_info[args.index_page]['hash'], config)
     loop = asyncio.get_event_loop()
-    loop.run_until_complete(check_tanner())
+    loop.run_until_complete(startup.check_tanner())
 
     pool = ProcessPoolExecutor(max_workers=multiprocessing.cpu_count())
     compare_version_fut = None
     if args.auto_update is True:
-        timeout = parse_timeout(args.update_timeout)
-        compare_version_fut = loop.run_in_executor(pool, compare_version_info, timeout)
+        timeout = startup.parse_timeout(args.update_timeout)
+        compare_version_fut = loop.run_in_executor(pool, startup.compare_version_info, timeout)
 
     if args.host_ip == 'localhost' and args.interface:
         args.host_ip = ni.ifaddresses(args.interface)[2][0]['addr']
 
     app = HttpRequestHandler(meta_info, args, snare_uuid, debug=args.debug, keep_alive=75)
-    drop_privileges()
+    startup.drop_privileges()
     print('serving with uuid {0}'.format(snare_uuid.decode('utf-8')))
     print("Debug logs will be stored in", log_debug)
     print("Error logs will be stored in", log_err)
